chore(EMS-1963): temporarily rename map CIS countries directory (git cache)
