+++ src/api/custom-resolvers/queries/get-APIM-CIS-countries/index.test.ts
@@ -1,6 +1,6 @@
 import getApimCisCountries from '.';
 import APIM from '../../../integrations/APIM';
-import mapCisCountries from '../../../helpers/map-CIS-countries';
+import mapCisCountries from '../../../helpers/map-CIS-countries-RENAMED';
 import mockCisCountriesResponse from '../../../test-mocks/mock-CIS-countries-response';
 import mockCisCountries from '../../../test-mocks/mock-CIS-countries';
