+++ src/api/helpers/map-CIS-countries-RENAMED/map-CIS-country/map-short-term-cover-available/index.ts
